+++ project-spring-boot/demo/src/main/java/com/example/demo/controller/user/ProjectController.java
@@ -5,6 +5,8 @@ import java.util.List;
 import javax.servlet.http.HttpSession;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.CrossOrigin;
@@ -43,8 +45,6 @@ public class ProjectController {
 	private MemberService memberService;
 	@Autowired
 	private RewardService rewardService;
-	@Autowired
-	private CommunityService communityService;
 	
 	//프로젝트 인덱스 페이지
 	@GetMapping("/index")
@@ -80,6 +80,17 @@ public class ProjectController {
 		return "/user/project/list";
 	}
 	
+	//프로젝트 좋아요 리스트 페이지
+		@GetMapping("/liked")
+		public String liked(Model model) {
+			log.info("project list");
+			
+			model.addAttribute("category", projectService.getCategory());
+			model.addAttribute("status", projectService.getStatus());
+			
+			return "/user/project/liked";
+		}
+	
 	//프로젝트 리스트 로드
 	@CrossOrigin(origins = "*", allowedHeaders = "*")
 	@ResponseBody
@@ -97,12 +108,31 @@ public class ProjectController {
 		return proejctAjaxListVo;
 	}
 	
+	//프로젝트 좋아요 리스트 로드
+	@CrossOrigin(origins = "*", allowedHeaders = "*")
+	@ResponseBody
+	@GetMapping("/likedajax")
+	public ProejctAjaxListVo getlikedAjax(ProjectFilteringVo filter, HttpSession session){
+		log.info("project liked");
+		
+		SessionMember member = (SessionMember) session.getAttribute("member");
+		filter.setMemberId(member.getId());
+		
+		ProejctAjaxListVo proejctAjaxListVo = new ProejctAjaxListVo();
+		
+		if(filter.getPage()==1) {
+			proejctAjaxListVo.setProjectCount(projectService.getLikedCount(filter));
+		}
+		proejctAjaxListVo.setProjectList(projectService.getLiked(filter));
+		
+		return proejctAjaxListVo;
+	}
+	
 	//프로젝트 좋아요 가져오기
 	@CrossOrigin(origins = "*", allowedHeaders = "*")
 	@ResponseBody
 	@GetMapping("/{projectId}/like")
-	public int getLike(@PathVariable("projectId") Integer projectId, HttpSession session) {
-		System.out.println("getLike");
+	public Integer getLike(@PathVariable("projectId") Integer projectId, HttpSession session) {
 		SessionMember member = (SessionMember) session.getAttribute("member");
 		VoteDto voteDto = new VoteDto();
 		
@@ -116,15 +146,14 @@ public class ProjectController {
 	@CrossOrigin(origins = "*", allowedHeaders = "*")
 	@ResponseBody
 	@PostMapping("/{projectId}/like")
-	public int updateLike(@PathVariable("projectId") Integer projectId, HttpSession session) {
-		System.out.println("updateLike");
+	public ResponseEntity<Integer> updateLike(@PathVariable("projectId") Integer projectId, HttpSession session) {
 		SessionMember member = (SessionMember) session.getAttribute("member");
+		if(member==null) return new ResponseEntity<Integer>(HttpStatus.UNAUTHORIZED);
 		VoteDto voteDto = new VoteDto();
-		
 		voteDto.setMemberId(member.getId());
 		voteDto.setProjectId(projectId);
 		
-		return projectService.updateLike(voteDto);
+		return new ResponseEntity<Integer>(projectService.updateLike(voteDto), HttpStatus.OK);
 	}
 	
 	//프로젝트 스토리 로드
@@ -137,96 +166,12 @@ public class ProjectController {
 		return projectService.getStory(id);
 	}
 	
-	//프로젝트 스토리 로드
+	//프로젝트 카테고리 로드
 	@CrossOrigin(origins = "*", allowedHeaders = "*")
 	@ResponseBody
 	@GetMapping("/category")
 	public List<ProjectCategoryDto> getCategory() {
 		
 		return projectService.getCategory();
-	}
-	
-	//프로젝트 커뮤니티 로드
-	@CrossOrigin(origins = "*", allowedHeaders = "*")
-	@ResponseBody
-	@GetMapping("/{id}/community")
-	public List<CommunityVo> getCommunity(@PathVariable("id") Integer id, @RequestParam("category") Integer category, Model model) {
-		log.info("project community : " + id);
-		log.info("category : " + category);
-		
-		return communityService.findByProjectId(id, category);
-	}
-	
-	//프로젝트 커뮤니티 작성
-	@CrossOrigin(origins = "*", allowedHeaders = "*")
-	@ResponseBody
-	@PostMapping("/{id}/community")
-	public void writeCommunity(@PathVariable("id") Integer id, CommunityDto communityDto, HttpSession session, Integer writerId, Integer parentId) {
-		SessionMember member = (SessionMember) session.getAttribute("member");
-		communityDto.setMemberId(member.getId());
-		communityDto.setProjectId(id);
-		communityDto.setCommunityCategoryId(writerId == communityDto.getMemberId() ? 2 : 1);
-		
-		System.out.println(parentId);
-		if(parentId==null) {
-			communityService.save(communityDto);
-		} else {
-			communityDto.setParentCommunityId(parentId);
-			communityService.saveReply(communityDto);
-		}
-		
-	}
-	
-	//프로젝트 커뮤니티 포스트 로드
-	@CrossOrigin(origins = "*", allowedHeaders = "*")
-	@ResponseBody
-	@GetMapping("/{projectId}/community/{postId}")
-	public CommunityVo readPost(@PathVariable("postId") Integer postId) {
-		log.info("readPost()");
-		
-		return communityService.findById(postId);
-	}
-	
-	//프로젝트 커뮤니티 포스트 댓글 로드
-	@CrossOrigin(origins = "*", allowedHeaders = "*")
-	@ResponseBody
-	@GetMapping("/{projectId}/community/{postId}/reply")
-	public List<CommunityVo> getPostReply(@PathVariable("postId") Integer parentId) {
-		log.info("getPostReply()");
-		
-		return communityService.findReply(parentId);
-	}
-	
-	//프로젝트 커뮤니티 포스트 개수 로드
-	@CrossOrigin(origins = "*", allowedHeaders = "*")
-	@ResponseBody
-	@GetMapping("/{projectId}/community/count")
-	public int getCount(@PathVariable("projectId") Integer projectId) {
-		log.info("getCount()");
-		
-		return communityService.getCount(projectId);
-	}
-	
-	//커뮤니티 업데이트
-	@CrossOrigin(origins = "*", allowedHeaders = "*")
-	@ResponseBody
-	@PutMapping("/community/{communityId}")
-	public Integer update(@PathVariable("communityId") Integer communityId, CommunityDto communityDto) {
-		log.info("community update()" + communityId);
-		
-		communityDto.setId(communityId);
-		communityService.update(communityDto);
-		
-		return communityDto.getId();
-	}
-	
-	//커뮤니티 삭제
-	@CrossOrigin(origins = "*", allowedHeaders = "*")
-	@ResponseBody
-	@DeleteMapping("/community/{communityId}")
-	public void delete(@PathVariable("communityId") Integer communityId) {
-		log.info("community delete()" + communityId);
-		
-		communityService.delete(communityId);
 	}
 }
