+++ project-spring-boot/demo/src/main/java/com/example/demo/domain/vo/ProjectFilteringVo.java
@@ -22,4 +22,6 @@ public class ProjectFilteringVo {
 	
 	private Integer sort;
 	private Integer page;
+	
+	private Integer memberId;
 }
