+++ project-spring-boot/demo/src/main/resources/mappers/projectMapper.xml
@@ -226,4 +226,40 @@
 			</where>
 	</select>
 	
+	<select id="findLiked" resultType="ProjectVo" parameterType="ProjectFilteringVo">
+		select * from (
+			select rownum rn, T.* from (
+				select P.* from vote V inner join (
+				select P.*, C.name project_category, nvl(RRR.sponsor, 0) sponsor, nvl(RRR.collected, 0) collected, nvl(round(collected/P.target_amount*100), 0) rate from (
+				        select * from project
+				    ) P 
+				    left outer join project_category C on P.project_category_id = C.id
+				    left outer join (select R.project_id, count(R.id) sponsor, sum(RR.reward_sum) collected from reserve R 
+				    left outer join reserve_reward RR on R.id = rr.reserve_id where R.reserve_status_category_id = 1 or R.reserve_status_category_id = 2 group by R.project_id)
+				    RRR on P.id = RRR.project_id
+				) P on V.project_id = P.id where V.member_id = #{memberId}
+		) T )
+		<where>
+			<if test="page != null">
+				and rn between #{page}*18-17 and #{page}*18
+			</if>
+			<if test="page == null">
+				and rn between 1 and 18
+			</if>
+		</where>
+	</select>
+	
+	<select id="getLikedCount" resultType="int" parameterType="ProjectFilteringVo">
+		select count(*) from (
+			select P.* from vote V inner join (
+			select P.*, C.name project_category, nvl(RRR.sponsor, 0) sponsor, nvl(RRR.collected, 0) collected, nvl(round(collected/P.target_amount*100), 0) rate from (
+			        select * from project
+			    ) P 
+			    left outer join project_category C on P.project_category_id = C.id
+			    left outer join (select R.project_id, count(R.id) sponsor, sum(RR.reward_sum) collected from reserve R 
+			    left outer join reserve_reward RR on R.id = rr.reserve_id where R.reserve_status_category_id = 1 or R.reserve_status_category_id = 2 group by R.project_id)
+			    RRR on P.id = RRR.project_id
+			) P on V.project_id = P.id where V.member_id = #{memberId}
+		)
+	</select>
 </mapper>
