+++ project-spring-boot/demo/src/main/webapp/WEB-INF/views/user/common/header.jsp
@@ -61,7 +61,7 @@
 	        </security:authorize>
             <li><a class="dropdown-item" href="#">메세지</a></li>
             <li><a class="dropdown-item" href="#">후원현황</a></li>
-            <li><a class="dropdown-item" href="#">좋아하는 프로젝트</a></li>
+            <li><a class="dropdown-item" href="/project/liked">좋아하는 프로젝트</a></li>
             <li><a class="dropdown-item" href="#">내가 만든 프로젝트</a></li>
             <li><a class="dropdown-item" href="/member">내 정보</a></li>
             <li><a class="dropdown-item" href="/member/edit/info">설정</a></li>
@@ -105,9 +105,9 @@
 					</div>
 					<div class="modalDivider"></div>
 					<div class="modelMenuItemContainer">
-						<div class="modelMenuItem"><a href="">모든 프로젝트</a></div>
-						<div class="modelMenuItem"><a href="">성공 임박 프로젝트</a></div>
-						<div class="modelMenuItem"><a href="">공개예정 프로젝트</a></div>
+						<div class="modelMenuItem"><a href="/project/list">모든 프로젝트</a></div>
+						<div class="modelMenuItem"><a href="/project/list">?</a></div>
+						<div class="modelMenuItem"><a href="/project/list">?</a></div>
 					</div>
 					<div class="modalDivider"></div>
 					<div class="modalCategoryContainer">
